# WalletService/WalletService.Api/Schema/openapi.yaml
openapi: 3.0.1
info:
  title: WalletService
  description: API for managing digital wallets
  version: 1.0.0
tags:
  - name: Wallets
paths:
  /api/wallets:
    post:
      summary: Create Wallet
      description: Creates a new wallet. Requires the Support role.
      operationId: createWallet
      tags:
        - Wallets
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletSchemaRequest'
        required: true
      responses:
        '200':
          description: Wallet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletSchemaIdResponse'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden – requires Support role
        '422':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error
      security:
        - BearerAuth: []
  /api/wallets/email/{email}:
    get:
      summary: Find Wallet by Email
      description: Returns a wallet by its email address.
      operationId: findWalletByEmail
      tags:
        - Wallets
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Wallet found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletSchemaResponse'
        '401':
          description: Unauthorized
        '404':
          description: Wallet not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error
      security:
        - BearerAuth: []
  /api/wallets/{walletId}:
    get:
      summary: Find Wallet
      description: Returns a wallet by its identifier.
      operationId: findWallet
      tags:
        - Wallets
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Wallet found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletSchemaResponse'
        '401':
          description: Unauthorized
        '404':
          description: Wallet not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error
      security:
        - BearerAuth: []
    patch:
      summary: Update Wallet
      description: Partially updates an existing wallet.
      operationId: updateWallet
      tags:
        - Wallets
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletPatchSchemaRequest'
        required: true
      responses:
        '200':
          description: Wallet updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletSchemaIdResponse'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
        '404':
          description: Wallet not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error
      security:
        - BearerAuth: []
    delete:
      summary: Delete Wallet
      description: Deletes a wallet by its identifier. Requires the Support role.
      operationId: deleteWallet
      tags:
        - Wallets
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Wallet deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden – requires Support role
        '404':
          description: Wallet not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error
      security:
        - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    WalletSchemaResponse:
      type: object
      properties:
        WalletId:
          type: string
          format: uuid
        Name:
          type: string
        LastName:
          type: string
        DocumentNumber:
          type: string
        DocumentType:
          type: string
        Currency:
          type: string
        Email:
          type: string
        Phone:
          type: string
        DailyLimit:
          type: number
          format: double
        balanceAmount:
          type: number
          format: double
      required:
        - WalletId
        - Name
        - LastName
        - DocumentNumber
        - DocumentType
        - Currency
        - Email
        - Phone
        - DailyLimit
        - balanceAmount
    WalletSchemaRequest:
      type: object
      properties:
        Name:
          type: string
        LastName:
          type: string
        DocumentNumber:
          type: string
        DocumentType:
          type: string
        Email:
          type: string
        Phone:
          type: string
        Currency:
          type: string
        DailyLimit:
          type: number
          format: double
      required:
        - Name
        - LastName
        - DocumentNumber
        - DocumentType
        - Email
        - Phone
        - Currency
        - DailyLimit
    WalletSchemaIdResponse:
      type: object
      properties:
        WalletId:
          type: string
          format: uuid
      required:
        - WalletId
    WalletPatchSchemaRequest:
      type: object
      properties:
        Name:
          type: string
          nullable: true
        LastName:
          type: string
          nullable: true
        DocumentNumber:
          type: string
          nullable: true
        DocumentType:
          type: string
          nullable: true
        Email:
          type: string
          nullable: true
        Phone:
          type: string
          nullable: true
        DailyLimit:
          type: number
          format: double
          nullable: true
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
  responses: {}
servers: []
security: []